User roles, court attribute, charts, and more

+++ app/controllers/lawsuits_controller.rb
@@ -2,7 +2,20 @@ class LawsuitsController < ApplicationController
   #before_action :set_lawsuit, only: [:show, :edit, :update, :destroy]
   def index
     @category = params[:category] || Lawsuit.categories.keys.first
-    @lawsuits = Lawsuit.where(category: @category).order(created_at: :desc)
+    @lawsuits = Lawsuit.important_lawsuits(@category)
+    # Apply filtering based on search parameters
+    if params[:plaintiff].present? || params[:lawsuit_number].present? || params[:status].present?
+      @lawsuits = Lawsuit.filter_by_params(@category, params)
+    end
+
+    @lawsuits = @lawsuits.paginate(page: params[:page], per_page: 10)
+
+
+    respond_to do |format|
+      format.html  # Renders the full page (not needed for Turbo requests)
+      format.turbo_stream  # Handles the Turbo Stream response
+    end
+
   end
 
   def show
@@ -33,6 +46,7 @@ class LawsuitsController < ApplicationController
 
   def update
     @lawsuit = Lawsuit.find(params[:id])
+    authorize @lawsuit
     if lawsuit_params[:pdf_files]
       @lawsuit.pdf_files.attach(lawsuit_params[:pdf_files])
     end
@@ -44,6 +58,7 @@ class LawsuitsController < ApplicationController
   end
   def destroy
     @lawsuit = Lawsuit.find(params[:id])
+    authorize @lawsuit
     @category = @lawsuit.category
     if @lawsuit.destroy
       flash[:notice] = "Lawsuit was successfully deleted."
@@ -57,7 +72,7 @@ class LawsuitsController < ApplicationController
 
   private
   def lawsuit_params
-    params.require(:lawsuit).permit(:title, :category, :status, :description, :context_type, :plaintiff, :lawsuit_claim, :lawsuit_number,
+    params.require(:lawsuit).permit(:title, :category, :status, :description, :context_type, :plaintiff, :lawsuit_claim, :lawsuit_number, :court,
                                     comments_attributes: [ :id, :content, :user_id ], pdf_files: [])
   end
 end
